# terraform/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Serverless Blog
...
paths:
...
  /api/post:
    get:
      summary: List posts
      operationId: listPosts
      responses:
        '200':
          description: Success request
          content:
            application/json:
               schema:
                 $ref: '#/components/schemas/ListPosts'
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: ${SERVERLESS_CONTAINER_ID}
    post:
      summary: Create post
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '201':
          description: Success creation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: ${SERVERLESS_CONTAINER_ID}
...
components:
  schemas:
...
    Post:
      title: Post
      properties:
        name:
          type: string
        text:
          type: string
        created_at:
          type: timestamp
    ListPosts:
      title: List posts
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Post'
    CreatePost:
      title: Create Post Request
      required:
        - name
        - text
      properties:
        name:
          type: string
        text:
          type: string